thitd

+++ src/main/java/breakingPointTool/calculations/FindSimilarArtifacts.java
@@ -173,7 +173,7 @@ public class FindSimilarArtifacts
         double Similarity = 0;
         if (metric1 != 0 && metric2 != 0) 
         {
-            Similarity = 100 - (Math.abs(metric1 - metric2) / (double) Math.max(metric1, metric2) * 100);
+            Similarity = 100 - (Math.abs(metric1 - metric2) / Math.max(metric1, metric2) * 100);
         }
         return Similarity;
     }
