+++ src/main/java/breakingPointTool/calculations/FindSimilarClasses.java
@@ -67,7 +67,7 @@ public class FindSimilarClasses
         double Similarity = 0;
         if (metric1 != 0 && metric2 != 0) 
         {
-            Similarity = 100 - (Math.abs(metric1 - metric2) / (double) Math.max(metric1, metric2) * 100);
+            Similarity = 100 - (Math.abs(metric1 - metric2) / Math.max(metric1, metric2) * 100);
         }
         return Similarity;
     }
